Remove api-mock, changes the code that depends on it.
I wrote a seperate express server for mocking HTTP request, but it turned
out such way is quite inconvenient. First I had to start it before started
the project, and when testing, everytime I wanted to chagne the mocking
data I had to restart it. Therefore, it is now removed, and wherever needs
a HTTP call, I wrote a async function that returns a dummy data.