Added Login component (empty) and the protected route logic to the project.
If a user hasn't logged in yet, it will redirect to the login page when they
try entering any page other than the login page. Similarity, if they has logged
in, it will redirect to the main page when they try entering the login page.

The user's info and login token are now just dummy data written in the appUser
redux slice.