Keep the button's badge visible until users enter then leave the notification page.
There is a button that will show the notification page, and there is a badge on it
that shows total of unread notifications. The badge would disappear when users entered
the page, yet users can still see which notification is new. To did this I had to
keep an extra state a.k.a `badgeNumber``. It makes the code confusing and doesn't improve
the UX a lot, so now it is deleted. The badge will shows so long as there are
notifications that haven't been read. When users close the notification page, we will
mark all notifications as read, and the badge will be hidden.