Removed the check mark next to a successfully sent message.
A checkmark whould be shown when a client sent a message to the server successfully
(when the server responsed OK). The problem is, at this moment, the message is only
garentee to reach to the server rather than the recipient. Showing a check mark on
a message may strongly imply that it has been sent to the recipient. Altough without
a check mark such implication still exists, but I think it is less strong.