Store chat messages on the server instead of in a local indexedDB database in the browser.
Influences:

Since we don't use indexedDB anymore, if the message fails to reach the server.
it will disappears like it hasn't been sent before when users refresh the app.

Refractorations:

emitEvent is renamed to emitSocketIOEvent. The parameters are changed and so do
the functionaility. With it now you can send multiple events to multiple users
in one call (combination of events and users).

New deps:

Added lodash to the server project.