Containerized the server into Docker.
Codes are adjusted to let them fit into the Docker's environment. This includes:

The settings file now can be a TypeScript ESM module that default exports an app's
setting. (JSON file is still supported).

It is necessary because you cannot change the settings file located in a container.
To let users to set/override app's settings, e.g. set the TLS certificate's
file path, the only way is let them set the environment variables, and inside
a container we assign the environment variable to the right setting's option.

A new CLI option `--settings-file` is added, so that we can use the development
setting when we `yarn dev`, and use the production setting when we run the app
in a Docker container.