Writing responsive UI (part 1).
On a small viewport (e.g. a mobile device), the app will only shows the
contact list (that is the sidebar) on the home page (path = "/"), and
will hide the sidebar on other pages. For example, on a message window
(path = /contact/<contactUserID>/chat), the app will only shows the
message window, and will hide the contact list.