The web app now resizes and crops selected avatar image before uploading it to the server.
Bug fixes:

Also fixed a bug that after I changed avatar the UI isn't updated to show the new avatar.
I have to press F5 and refresh the page to see the new avatar.

The reason is because the URL to the avatar will not change over time, for example, if
a user whose user ID is 1, his/her's avatar's URL will always be `/userdata/1/user-avatar`,
though the content the URL points to will be updated.

The solution is simple, we now add a dummy querystring to the avatar's URL whenever users
update their avatar.